requirements.txt
langchain_community
tiktoken
langchain-openai
langchainhub
chromadb
langchain
python-dotenv
beautifulsoup4
langchain-groq
youtube-transcript-api
pytube
ragatouille
langgraph